+++ Engine/Sprite.h
@@ -22,9 +22,9 @@ class Sprite
 		XMVECTOR uv;
 	};
 protected:
-	UINT64 vertexNum_;
+	UINT vertexNum_;
 	VERTEX* vertices_;
-	UINT64 indexNum_;//インデックス情報
+	UINT indexNum_;//インデックス情報
 	int* index_;//インデックス情報
 	ID3D11Buffer* pVertexBuffer_;	//頂点バッファ
 	ID3D11Buffer* pIndexBuffer_;   //インデックスバッファ
